+++ src/components/content/MessageWithPopper.tsx
@@ -1,13 +1,5 @@
-import {
-    Box,
-    Button,
-    ClickAwayListener,
-    Fade,
-    Popper,
-    useMediaQuery,
-    useTheme,
-} from '@mui/material';
-import { logDialogBackground, popperIndex } from 'context/Theme';
+import { Box, Button } from '@mui/material';
+import PopperWrapper from 'components/shared/PopperWrapper';
 import { MouseEvent, ReactNode, useState } from 'react';
 import { FormattedMessage } from 'react-intl';
 
@@ -17,23 +9,12 @@ interface Props {
 }
 
 function MessageWithPopper({ messageId, popper }: Props) {
-    const theme = useTheme();
-    const belowMd = useMediaQuery(theme.breakpoints.down('md'));
-
     const [open, setOpen] = useState<boolean>(false);
     const [anchorEl, setAnchorEl] = useState<null | HTMLElement>(null);
 
-    const canBeOpen = open && Boolean(anchorEl);
-    const id = canBeOpen ? 'transition-popper' : undefined;
-
-    const handlers = {
-        togglePopper: (event: MouseEvent<HTMLElement>) => {
-            setAnchorEl(event.currentTarget);
-            setOpen((previousOpen) => !previousOpen);
-        },
-        externalAreaClicked: () => {
-            setOpen(false);
-        },
+    const togglePopper = (event: MouseEvent<HTMLElement>) => {
+        setAnchorEl(event.currentTarget);
+        setOpen((previousOpen) => !previousOpen);
     };
 
     return (
@@ -46,41 +27,19 @@ function MessageWithPopper({ messageId, popper }: Props) {
                         <Button
                             variant="text"
                             size="small"
-                            onClick={handlers.togglePopper}
+                            onClick={togglePopper}
                             sx={{ mb: 0.5 }}
                         >
                             <FormattedMessage id={`${messageId}.buttonLabel`} />
                         </Button>
 
-                        <Popper
-                            id={id}
-                            open={open}
+                        <PopperWrapper
                             anchorEl={anchorEl}
-                            transition
-                            sx={{ zIndex: popperIndex }}
+                            open={open}
+                            setOpen={setOpen}
                         >
-                            {({ TransitionProps }) => (
-                                <ClickAwayListener
-                                    onClickAway={handlers.externalAreaClicked}
-                                >
-                                    <Fade {...TransitionProps} timeout={350}>
-                                        <Box
-                                            sx={{
-                                                maxWidth: belowMd ? 450 : 650,
-                                                p: 2,
-                                                borderRadius: 5,
-                                                bgcolor:
-                                                    logDialogBackground[
-                                                        theme.palette.mode
-                                                    ],
-                                            }}
-                                        >
-                                            {popper}
-                                        </Box>
-                                    </Fade>
-                                </ClickAwayListener>
-                            )}
-                        </Popper>
+                            {popper}
+                        </PopperWrapper>
                     </>
                 ),
             }}
