+++ src/components/shared/ButtonWithPopper.tsx
@@ -1,13 +1,5 @@
-import {
-    Box,
-    Button,
-    ClickAwayListener,
-    Fade,
-    Popper,
-    useMediaQuery,
-    useTheme,
-} from '@mui/material';
-import { logDialogBackground, popperIndex } from 'context/Theme';
+import { Button } from '@mui/material';
+import PopperWrapper from 'components/shared/PopperWrapper';
 import { MouseEvent, ReactNode, useState } from 'react';
 import { FormattedMessage } from 'react-intl';
 
@@ -18,63 +10,27 @@ interface Props {
 }
 
 function ButtonWithPopper({ messageId, popper, startIcon }: Props) {
-    const theme = useTheme();
-    const belowMd = useMediaQuery(theme.breakpoints.down('md'));
-
     const [open, setOpen] = useState<boolean>(false);
     const [anchorEl, setAnchorEl] = useState<null | HTMLElement>(null);
 
-    const canBeOpen = open && Boolean(anchorEl);
-    const id = canBeOpen ? 'transition-popper' : undefined;
-
-    const handlers = {
-        togglePopper: (event: MouseEvent<HTMLElement>) => {
-            setAnchorEl(event.currentTarget);
-            setOpen((previousOpen) => !previousOpen);
-        },
-        externalAreaClicked: () => {
-            setOpen(false);
-        },
+    const togglePopper = (event: MouseEvent<HTMLElement>) => {
+        setAnchorEl(event.currentTarget);
+        setOpen((previousOpen) => !previousOpen);
     };
 
     return (
         <>
             <Button
                 startIcon={startIcon}
-                onClick={handlers.togglePopper}
+                onClick={togglePopper}
                 sx={{ mb: 0.5 }}
             >
                 <FormattedMessage id={messageId} />
             </Button>
 
-            <Popper
-                id={id}
-                open={open}
-                anchorEl={anchorEl}
-                transition
-                sx={{ zIndex: popperIndex }}
-            >
-                {({ TransitionProps }) => (
-                    <ClickAwayListener
-                        onClickAway={handlers.externalAreaClicked}
-                    >
-                        <Fade {...TransitionProps} timeout={350}>
-                            <Box
-                                sx={{
-                                    maxWidth: belowMd ? 450 : 650,
-                                    p: 2,
-                                    borderRadius: 5,
-                                    boxShadow: 2,
-                                    bgcolor:
-                                        logDialogBackground[theme.palette.mode],
-                                }}
-                            >
-                                {popper}
-                            </Box>
-                        </Fade>
-                    </ClickAwayListener>
-                )}
-            </Popper>
+            <PopperWrapper anchorEl={anchorEl} open={open} setOpen={setOpen}>
+                {popper}
+            </PopperWrapper>
         </>
     );
 }
